requirements.txt
flet==0.80.3
psycopg2-binary
pytest
